Extract `Axes` and `Anchor` from `ExtendedEntity.kt`
