+++ src/com/reco1l/andengine/Axes.kt
@@ -1,0 +1,27 @@
+package com.reco1l.andengine
+
+/**
+ * The axes in a bi-dimensional space.
+ */
+enum class Axes {
+
+    /**
+     * Horizontal axis.
+     */
+    X,
+
+    /**
+     * Vertical axis.
+     */
+    Y,
+
+    /**
+     * Both axes.
+     */
+    Both,
+
+    /**
+     * None of the axes.
+     */
+    None
+}
