+++ src/com/reco1l/andengine/ExtendedEntity.kt
@@ -405,48 +405,3 @@ abstract class ExtendedEntity(
 }
 
 
-enum class Anchor(val factorX: Float, val factorY: Float) {
-
-    TopLeft(0f, 0f),
-
-    TopCenter(0.5f, 0f),
-
-    TopRight(1f, 0f),
-
-    CenterLeft(0f, 0.5f),
-
-    Center(0.5f, 0.5f),
-
-    CenterRight(1f, 0.5f),
-
-    BottomLeft(0f, 1f),
-
-    BottomCenter(0.5f, 1f),
-
-    BottomRight(1f, 1f)
-
-}
-
-
-enum class Axes {
-
-    /**
-     * The entity will automatically adjust its size to the width of the parent.
-     */
-    X,
-
-    /**
-     * The entity will automatically adjust its size to the height of the parent.
-     */
-    Y,
-
-    /**
-     * The entity will automatically adjust its size to the width and height of the parent.
-     */
-    Both,
-
-    /**
-     * The entity will not automatically adjust its size.
-     */
-    None
-}
